+++ lib/rule_container.dart
@@ -1,0 +1,9 @@
+import 'validators/base_validator.dart';
+
+class RuleContainer {
+  List<BaseValidator> rules = <BaseValidator>[];
+
+  Function getter;
+
+  RuleContainer(this.getter);
+}