+++ lib/src/message.dart
@@ -1,7 +1,0 @@
-mixin Message {
-  String _message;
-  String get message => _message;
-  set message(String s) {
-    _message = s;
-  }
-}