+++ lib/src/validationFailure.dart
@@ -1,4 +1,0 @@
-class ValidationFailure {
-  String associatedWith;
-  String errorMessage;
-}