+++ lib/validation_failure.dart
@@ -1,0 +1,4 @@
+class ValidationFailure {
+  String? associatedWith;
+  String? errorMessage;
+}