+++ CHANGELOG.md
@@ -1,7 +1,14 @@
+# Change logs
+
 ## 1.0.0 Initial Version
 
 - Fluent validation library released.
 
 ## 1.0.1 Update
+
 - Library update
-- Changed validationResult.errorText to return null instead of empty string if no validation errors exist. 
+- Changed validationResult.errorText to return null instead of empty string if no validation errors exist.
+
+## 1.0.2 Update
+
+- Null-Safety migration.